api/pyproject.toml
[project]
name = "experiment-queue-api"
version = "0.1.0"
description = "Experiment queue management API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic-settings>=2.7.0",
    "pyairtable>=3.0.0",
    "boto3>=1.35.0",
    "pandas>=2.2.0",
    "matplotlib>=3.9.0",
    "pillow>=11.0.0",
    "opencv-python-headless>=4.10.0",
    "google-api-python-client>=2.150.0",
    "google-auth>=2.36.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "pytz>=2024.2",
    "requests>=2.32.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "pyright>=1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "basic"
